+++ src/bibleTools/types.nim
@@ -1,0 +1,3 @@
+type
+  AvailableLanguages* = enum
+    ALDefault, ALEnglish, ALPortuguese